+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.x.x
+
+### Changes and Fixes
+
+- Module [SIREN](./docs/siren/Siren.md)
+    - Added support for multi-channel files on Re-pitch and Crossfade-loop
+    - Fixed crash when playing multi-channel files
+
 ## 2.5.0
 
 ### New modules
@@ -18,6 +26,8 @@
     - Fixed saving of SysEx data logging setting
 - Module [PANIC ROOM](./docs/panicroom/PanicRoom.md)
     - Added options to limit the number of allowed modules and cables
+- Module [STRIP](./docs/strip/Strip.md)
+    - Fixed crash when removing modules added by STRIP in plugin-version of Rack
 - Module [TRANSIT](./docs/transit/Transit.md)
     - Fade CV input is now additive to per-slot fade time (previously CV was only additive to the global _FADE_ knob)
     - Added Output-mode "Tipsy" for sending the snapshot text label (for modules with Tipsy-support like [TTY](https://library.vcvrack.com/StochasticTelegraph/TTY))